refactor: common폴더 위치 수정
