+++ src/app/page.tsx
@@ -3,7 +3,7 @@
 import styles from "./page.module.css";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
-import Button from "./common/Button/Button";
+import Button from "./components/common/Button/Button";
 
 export default function Home() {
   const router = useRouter();
